Split OpenTPL client into a separate class for separate tt_master and tt_meteo connections.
